fixed compile error
<==== Mercurial data ====>
Branch: dev_3.0.0

+++ nx_cloud/libcloud_db/src/ec2/dao/rdb/transaction_data_object.cpp
@@ -87,8 +87,8 @@ nx::db::DBResult TransactionDataObject::fetchTransactionsOfAPeerQuery(
     fetchTransactionsOfAPeerQuery.addBindValue(QLatin1String(systemId));
     fetchTransactionsOfAPeerQuery.addBindValue(peerId);
     fetchTransactionsOfAPeerQuery.addBindValue(dbInstanceId);
-    fetchTransactionsOfAPeerQuery.addBindValue(minSequence);
-    fetchTransactionsOfAPeerQuery.addBindValue(maxSequence);
+    fetchTransactionsOfAPeerQuery.addBindValue((qint64)minSequence);
+    fetchTransactionsOfAPeerQuery.addBindValue((qint64)maxSequence);
     if (!fetchTransactionsOfAPeerQuery.exec())
     {
         NX_LOGX(QnLog::EC2_TRAN_LOG,
